+++ GestionBoutiqueFlorale/Fournisseur.cs
@@ -8,6 +8,6 @@ namespace GestionBoutiqueFlorale
 {
     public class Fournisseur : Utilisateur
     {
-        public Fournisseur(string nom, string prenom, int telephone, string adresse) : base(nom, prenom, telephone, adresse) { }
+        public Fournisseur(string nom, string prenom, Int64 telephone, string adresse) : base(nom, prenom, telephone, adresse) { }
     }
 }
