+++ GestionBoutiqueFlorale/Vendeur.cs
@@ -8,7 +8,7 @@ namespace GestionBoutiqueFlorale
 {
     public class Vendeur : Utilisateur
     {
-        public Vendeur(string nom, string prenom, int telephone, string adresse) : base(nom, prenom, telephone, adresse) { }
+        public Vendeur(string nom, string prenom, Int64 telephone, string adresse) : base(nom, prenom, telephone, adresse) { }
        
 
     }
